+++ tests/bluetooth/audio/ascs/src/main.c
@@ -35,37 +35,14 @@
 
 DEFINE_FFF_GLOBALS;
 
-static struct bt_codec lc3_codec =
-	BT_CODEC_LC3(BT_CODEC_LC3_FREQ_ANY, BT_CODEC_LC3_DURATION_10,
-		     BT_CODEC_LC3_CHAN_COUNT_SUPPORT(1), 40u, 120u, 1u,
-		     (BT_AUDIO_CONTEXT_TYPE_CONVERSATIONAL | BT_AUDIO_CONTEXT_TYPE_MEDIA));
-
-static void pacs_cap_foreach_custom_fake(enum bt_audio_dir dir, bt_pacs_cap_foreach_func_t func,
-					 void *user_data)
-{
-	static const struct bt_pacs_cap cap[] = {
-		{
-			&lc3_codec,
-		},
-	};
-
-	for (size_t i = 0; i < ARRAY_SIZE(cap); i++) {
-		if (func(&cap[i], user_data) == false) {
-			break;
-		}
-	}
-}
-
 static void mock_init_rule_before(const struct ztest_unit_test *test, void *fixture)
 {
 	mock_bap_unicast_server_init();
 	mock_bt_iso_init();
 	mock_kernel_init();
-	PACS_FFF_FAKES_LIST(RESET_FAKE);
+	mock_bt_pacs_init();
 	mock_bap_stream_init();
 	mock_bt_gatt_init();
-
-	bt_pacs_cap_foreach_fake.custom_fake = pacs_cap_foreach_custom_fake;
 }
 
 static void mock_destroy_rule_after(const struct ztest_unit_test *test, void *fixture)
@@ -73,6 +50,7 @@ static void mock_destroy_rule_after(const struct ztest_unit_test *test, void *fi
 	mock_bap_unicast_server_cleanup();
 	mock_bt_iso_cleanup();
 	mock_kernel_cleanup();
+	mock_bt_pacs_cleanup();
 	mock_bap_stream_cleanup();
 	mock_bt_gatt_cleanup();
 }
