+++ tests/bluetooth/audio/mocks/include/pacs.h
@@ -10,9 +10,8 @@
 #include <zephyr/fff.h>
 #include <zephyr/bluetooth/audio/pacs.h>
 
-/* List of fakes used by this unit tester */
-#define PACS_FFF_FAKES_LIST(FAKE)                                                                  \
-	FAKE(bt_pacs_cap_foreach)                                                                  \
+void mock_bt_pacs_init(void);
+void mock_bt_pacs_cleanup(void);
 
 DECLARE_FAKE_VOID_FUNC(bt_pacs_cap_foreach, enum bt_audio_dir, bt_pacs_cap_foreach_func_t, void *);
